updating routing flow

+++ BougieCandles/ViewModels/ShoppingCartViewModel.cs
@@ -9,6 +9,6 @@ namespace BougieCandles.ViewModels
     public class ShoppingCartViewModel
     {
         public ShoppingCart ShoppingCart { get; set; }
-        public decimal ShoppingCartTotal { get; set; }
+        public double ShoppingCartTotal { get; set; }
     }
 }
